Loads JSON and saves poster.
Makes use of json2.js, which pretty darn neat.

Loads the JSON file, and runs a loop. 
On each loop, it reads the information present in the JSON file, and decides which layer to make visible, depending on the writeup type, specified in the JSON file. 

Convention: 
1: Shortie
2: Scribble
3: Poem
4: Story

After it has added the content from the JSON file to the selected layers, it (function processWriteup) runs the saveJPEG function to save the current canvas in JPEG, in a child folder named 'Output'.
Thereafter, it proceeds to make all layers invisible, and patiently waits for the next loop iteration in the main function, to generate another writeup.